Decouple activation feedback from window management
Window management and activation feedback are two unrelated problems. If
an application wants just to manage windows, sending activation objects
to it will be pointless. Same goes with just listening for activations, if
an app just wants to monitor activations, sending stacking order or
other properties to it will be pointless.

Since window management and activation feedback are two sort of
orthogonal problems, this change moves activation feedback out of the
management interface in a separate interface.