Make explicit that kde-output-management-v2 is onetime only
It's already onetime, meaning that the apply request will reset the
state of the output configuration object, but it will be easier to
implement this protocol on the compositor side if the apply request is
onetime only.

It would be also nice to mark applied and failed events as destructors,
but that would require bumping required wayland version and it means
bumping v2 to v3.