output device, output management: add brightness setting
This allows the user to set a brightness multiplier for all content,
including HDR content, and can be used to implement a convenient
brightness slider.
The setting is only supported in HDR mode right now, but it can be
extended later to apply to outputs in SDR mode as well.