fake-input: add keysym request
In x-d-p-kde we need to handle keysyms in addition to keycodes for remote desktop

While in theory we can infer the keycode for a given keysym it doesn't work reliably
in practice. This is because we need to take into account the current keyboard
layout. In principle we get this from wayland, but since x-d-p-kde is a background
process it doesn't usually receive any keyboard events from the compositor.

The compositor is in a much better position to decide how to process a keysym,
so let x-d-p-kde pass it to the compositor directly.

CCBUG: 489021