kde-output-{device,management}: Add ddc_ci_allowed
The Wayland event and request respectively are available starting
with v13 of each protocol.

In addition to the request and event, this commit also adds a
capability to indicate whether an output supports DDC/CI in the
first place. The request set_uses_ddc_ci recently added to the
kde-external-brightness-v1 protocol can be used to determine this
information for external brightness devices.