kde-external-brightness: Add set_uses_ddc_ci()
A new request to tell the compositor whether this brightness device
is using DDC/CI to apply its settings to the display.

This information can be used by the compositor to ignore the
brightness device if it exposes monitor issues. It can also be used
to prefer software brightness for frequent actions such as dimming,
maximizing monitor longevity.