appmenu: Add available event
The available event communicates to the client whether there is an
available appmenu and whether it can hide its menubar.

Currently, one way to determine whether the global menu is available is
to check whether the com.canonical.AppMenu.Registrar service name
is registered, but for some applications, it's undesired to make dbus
queries on startup, and in general, it would be nice to stop dragging
this X11 specific implementation detail for Wayland clients.