output device, -management: add pixel encoding setting
This allows the user to select a specific RGB or YCbCr encoding, instead of the
driver defaults. Usually this is not necessary, but some screens don't handle
all encodings properly, and it allows the user to make sure that they're not using
chroma subsampling when undesired (like on a computer monitor).